Refactor ParsedStringWithIndent into it's own file.
